+++ externs.js
@@ -1,0 +1,16 @@
+/**
+ * @fileoverview Externs for endianness 7.0
+ *
+ * @see https://github.com/rochars/endianness
+ * @externs
+ */
+
+/**
+ * Swap the byte ordering in a buffer. The buffer is modified in place.
+ * @param {!Array<number|string>|!Uint8Array} bytes The bytes.
+ * @param {number} offset The byte offset.
+ * @param {number=} start The start index. Assumes 0.
+ * @param {?number=} end The end index. Assumes the buffer length.
+ * @throws {Error} If the buffer length is not valid.
+ */
+function endianness(bytes, offset, start=0, end=null) {}