+++ index.d.ts
@@ -1,0 +1,6 @@
+// Type definitions for endianness 7.0
+// Project: https://github.com/rochars/endianness
+// Definitions by: Rafael S. Rocha <https://github.com/rochars>
+// Definitions: https://github.com/rochars/endianness
+
+export default function endianness(bytes: Array<number|string>|Uint8Array, offset?: number, start?: number, end?: number): void;